+++ lib/puppet/provider/cfnetwork_firewall_iface/firehol.rb
@@ -1,4 +1,6 @@
-require 'cffirehol/providerbase'
+
+# Done this way due to some weird behavior in tests also ignoring $LOAD_PATH
+require File.expand_path( '../../../../cffirehol/providerbase', __FILE__ )
 
 Puppet::Type.type(:cfnetwork_firewall_iface).provide(
     :firehol,