+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.9.4]
+
+- Added check verify that port ifaces are defined instead of not understandable error
+- Changed to require ruby modules by absolute path due to strange issues with $LOAD_PATH in some deployments
+
+## [0.9.3]
+
+- No changes, accident release.
+
 ## [0.9.2]
 
 - Fixed src/dst property processing with DHCP interfaces
@@ -25,6 +34,8 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial release
 
+[0.9.4](https://github.com/codingfuture/puppet-cffirehol/releases/tag/v0.9.4)
+[0.9.3](https://github.com/codingfuture/puppet-cffirehol/releases/tag/v0.9.3)
 [0.9.2](https://github.com/codingfuture/puppet-cffirehol/releases/tag/v0.9.2)
 [0.9.1](https://github.com/codingfuture/puppet-cffirehol/releases/tag/v0.9.1)
 [0.9.0](https://github.com/codingfuture/puppet-cffirehol/releases/tag/v0.9.0)